refactor: 챌린지 컴포넌트로 분리 후 동적 import
